extends files and move images to helper
